requirements.txt
langchain==0.0.117
llama-index==0.4.33
openai==0.27.2
python-dotenv==1.0.0
python-telegram-bot==20.1

